requirements.txt
Django>=4.0
djangorestframework
django-ipware
django-ratelimit
requests
celery
redis